修复新增bug

+++ src/main/resources/db/schema.sql
@@ -1,6 +1,6 @@
 drop table if exists `user`;
 create table `user`(
-id int(32) AUTO_INCREMENT primary key,
+id bigint(20) AUTO_INCREMENT primary key,
 `name` varchar(50),
 birthday date,
 gender int(1)
